+++ src/chevre.ts
@@ -1,4 +1,0 @@
-/**
- * Chevre API Client
- */
-export * from '@chevre/api-nodejs-client';