+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- 予約集計処理をCinerino使用へ変更
+
 ### Deprecated
 
 ### Removed